+++ Sources/NablaCore/Domain/Repositories/DeviceRepository.swift
@@ -19,32 +19,40 @@ final class DeviceRepositoryImpl: DeviceRepository {
                 switch result {
                 case let .failure(error):
                     logger.error(message: "Failed to register device", extra: ["error": error])
-                case let .success(deviceId):
-                    deviceLocalDataSource.deviceId = deviceId
-                    logger.info(message: "Registered device", extra: ["id": deviceId])
+                case let .success(remoteDevice):
+                    deviceLocalDataSource.deviceId = remoteDevice.deviceId
+                    if let sentry = remoteDevice.sentry {
+                        self.errorReporter.enable(dsn: sentry.dsn, env: sentry.env)
+                    } else {
+                        self.errorReporter.disable()
+                    }
+                    logger.info(message: "Registered device", extra: ["id": remoteDevice.deviceId])
                 }
             }
         )
     }
-    
+
     // MARK: Init
-    
+
     init(
         deviceLocalDataSource: DeviceLocalDataSource,
         deviceRemoteDataSource: DeviceRemoteDataSource,
-        logger: Logger
+        logger: Logger,
+        errorReporter: ErrorReporter
     ) {
         self.deviceLocalDataSource = deviceLocalDataSource
         self.deviceRemoteDataSource = deviceRemoteDataSource
         self.logger = logger
+        self.errorReporter = errorReporter
     }
-    
+
     // MARK: - Private
-    
+
     private let deviceLocalDataSource: DeviceLocalDataSource
     private let deviceRemoteDataSource: DeviceRemoteDataSource
     private let logger: Logger
-    
+    private let errorReporter: ErrorReporter
+
     private func serialize(_ module: Module) -> GQL.SdkModule {
         if module is MessagingModule {
             return .messaging